requirements.txt
# Core Dependencies
mcp>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-ollama>=0.1.0
fastmcp>=0.1.0
playwright>=1.40.0

# Development Dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
